[Feature] Edit products

+++ public/js/manageFactors.js
@@ -1,6 +1,13 @@
 const deleteFactor = async (e) => {
   const fid = e.target.getAttribute('data-factor-id');
 
+  const userConfirmed = await showModalConfirmation();
+
+  if (!userConfirmed) {
+    alert('Eliminación cancelada');
+    return;
+  }
+
   try {
     const response = await fetch(`/api/factors/${fid}`, {
       method: 'DELETE',
@@ -9,13 +16,29 @@ const deleteFactor = async (e) => {
       },
     });
     if (response.ok) {
-      alert('Factor deleted successfully');
+      alert('Factor eliminado correctamente');
       window.location.reload();
     } else {
       const error = await response.json();
-      alert(error.error.message);
+      alert('Factor no eliminado');
     }
   } catch (error) {
-    alert(error);
+    alert('Error al eliminar factor');
   }
+};
+const showModalConfirmation = () => {
+  return new Promise((resolve) => {
+    const modal = document.getElementById('confirmationModal');
+    modal.classList.remove('hidden');
+
+    document.getElementById('confirmBtn').onclick = () => {
+      modal.classList.add('hidden');
+      resolve(true);
+    };
+
+    document.getElementById('cancelBtn').onclick = () => {
+      modal.classList.add('hidden');
+      resolve(false);
+    };
+  });
 };
